+++ fe/src/pages/SignUpPage/subPages/NicknameSubPage.tsx
@@ -1,7 +1,7 @@
 import { postNicknameDuplicateCheck } from "@api/auth";
 import { HTTPSTATUS } from "@api/types";
 import useText from "@hooks/useText";
-import { validateNickname } from "@utils/textValidators";
+import { validateNickname } from "@utils/authInputValidators";
 import axios from "axios";
 import { ChangeEvent, useState } from "react";
 import SubPage from "./SubPage";
