+++ fe/src/utils/authInputValidators.ts
